김승곤 : paid 페이징처리 제외

+++ serial_novel_project/src/main/java/com/novel/paid/domain/Criteria.java
@@ -4,14 +4,14 @@ import lombok.Data;
 
 @Data
 public class Criteria {
-	private int pageNum = 1;
-	private int number = 10;
+	private long pageNum = 1;
+	private long number = 10;
 	
-	public int getPageStart() {
+	public long getPageStart() {
 		return(this.pageNum - 1) * number;
 	}
 	
-	public int getPageEnd() {
+	public long getPageEnd() {
 		return(this.pageNum * number);
 	}
 }
